# shared/openapi.yaml
openapi: 3.1.0
info:
  title: 'Web-block openapi'
  license:
    name: MIT
    identifier: MIT
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/'
    description: 'Backend server'
paths:
  /api/login:
    post:
      tags:
        - auth
      summary: 'Handle an incoming authentication request.'
      operationId: bc76a9d52929cab7a147ca7b5c527430
      requestBody:
        description: 'Login the user'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: 'User successfully logged in and token issued.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: 'Error with login credentials.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
        '422':
          description: 'Validation error.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
  /api/logout:
    delete:
      tags:
        - auth
      summary: 'Destroy an authenticated session.'
      operationId: 250045954f07adf40103e33a04e5d715
      responses:
        '204':
          description: 'User successfully logged out.'
      security:
        -
          scalar: sessionAuth
...
  /api/register:
    post:
      tags:
        - auth
      summary: 'Handle an incoming registration request.'
      operationId: apiRegister
      requestBody:
        description: 'Create user model and login in user.'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 'User successfully registered.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: 'Error with registering.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
...
  /api/projects:
    get:
      tags:
        - project
      summary: 'Display a listing of the resource.'
      operationId: 8d4a13ef3fc773ff4031e515460919ea
      responses:
        '200':
          description: test
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    post:
      tags:
        - project
      summary: 'Store a newly created resource in storage.'
      operationId: a4dd4c17bf5fc24c224357f03dcc848a
      requestBody:
        description: 'Name to crate the project'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreProjectRequest'
      responses:
        '201':
          description: 'Project created successfully'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string }
                type: object
        '400':
          description: 'Invalid url'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  /api/projects/create:
    post:
      tags:
        - project
      summary: 'Store a newly created resource in storage.'
      operationId: 5dae9d6654a488de0a9e7fc0a8800669
      requestBody:
        description: 'Project data'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreClientProjectRequest'
      responses:
        '204':
          description: 'Project created successfully'
        '400':
          description: 'Invalid data.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorObject'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
  '/api/projects/{project}':
    get:
      tags:
        - project
      summary: 'Display the specified resource.'
      operationId: 1352e06e55f6ab10d3702509cd385918
      parameters:
        -
          name: project
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
      security:
        -
          scalar: sessionAuth
    put:
      tags:
        - project
      summary: 'Update the specified resource in storage.'
      operationId: a93b65e3037bfd5ea5cdc1b3dd11ba53
      parameters:
        -
          name: project
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: 'Date to update project.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '204':
          description: 'Project updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
    delete:
      tags:
        - project
      summary: 'Remove the specified resource from storage.'
      operationId: 63c289a907902132dd95dd7ecb50143f
      parameters:
        -
          name: project
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 'Project deleted.'
      security:
        -
          scalar: sessionAuth
    patch:
      tags:
        - project
      operationId: ee35632abc0168537257bd8f5baa9f30
      parameters:
        -
          name: project
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: 'Date to update project.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '204':
          description: 'Project updated.'
        '401':
          description: Unauthenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorObject'
      security:
        -
          scalar: sessionAuth
...
components:
  schemas:
    ErrorObject:
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: 'Validation failed.'
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              example: 'Field is required'
      type: object
...
    AuthTokenResponse:
      required:
        - message
        - token
      properties:
        message:
          type: string
          example: 'Login successful'
        token:
          description: 'Sanctum plain text API token'
          type: string
          example: 1|abcdefghijklmnopqrstuvwxyzabcdefg
      type: object
    RegisterRequest:
      required:
        - name
      allOf:
        -
          properties:
            name:
              type: string
          type: object
        -
          $ref: '#/components/schemas/LoginRequest'
...
    LoginRequest:
      required:
        - email
        - password
        - remember
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        remember:
          type: boolean
          default: true
      type: object
...
    StoreClientProjectRequest:
      required:
        - id
        - name
        - description
        - x
        - 'y'
        - zoom
        - default_branch
        - oid
        - cwd
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        x:
          type: integer
        'y':
          type: integer
        zoom:
          type: number
          format: float
        default_branch:
          type: string
        url:
          type: string
        oid:
          type: string
        cwd:
          type: string
      type: object
    StoreProjectRequest:
      required:
        - nameOrUrl
      properties:
        nameOrUrl:
          type: string
      type: object
...
    UpdateProjectRequest:
      required:
        - name
        - description
        - x
        - 'y'
        - zoom
      properties:
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        x:
          type: integer
        'y':
          type: integer
        zoom:
          type: number
          format: float
      type: object
...
    Project:
      required:
        - id
        - name
        - description
        - x
        - 'y'
        - zoom
        - default_branch
        - url
        - oid
        - cwd
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        x:
          type: integer
          default: 0
        'y':
          type: integer
          default: 0
        zoom:
          type: number
          format: float
          default: 1
        default_branch:
          type: string
          default: main
        url:
          type: string
        oid:
          type: string
        cwd:
          type: string
          default: /
      type: object
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      description: "This API uses session-based authentication. The session cookie is automatically handled by the \n            middleware, and clients do not need to include it in their requests."
      name: web_block_session
      in: cookie
      flows:
        implicit:
          authorizationUrl: /api/auth/redirect
          scopes:
            - read
            - write
...